Move app component folders.
