+++ src/app/components/slider.js
