+++ src/app/components/switcher/index.css
