+++ src/app/components/switcher/index.js
