+++ src/app/index.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
 import styled from 'react-emotion';
 
-import Switcher from './routes/common/components/switcher/index';
-import ServiceWorker from './routes/common/components/serviceWorker';
+import Switcher from './components/switcher/index';
+import ServiceWorker from './components/serviceWorker';
 import getRoute from './routes';
 import routes from './routesMap';
